third_party/blink/web_tests/wpt_internal/html/canvas/placeElement/placeElement-hit-testing.html

<!DOCTYPE html>

<head>
  <title>CanvasRenderingContext2D.PlaceElement(): Placed elements' hit testing works.</title>
  <link rel="author" title="Aaron Krajeski" href="mailto:aaronhk@chromium.org" />
  <link rel="help" href="https://github.com/WICG/canvas-place-element" />
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/resources/testdriver.js"></script>
  <script src="/resources/testdriver-actions.js"></script>
  <script src="/resources/testdriver-vendor.js"></script>

  <style>
    * {
      margin: 0;
      --magenta: rgb(255, 0, 255);
      --brightGreen: rgb(0, 255, 0);
      --width: 20px;
      --height: 10px;
    }

    #fallback {
      background-color: var(--brightGreen);
      width: var(--width);
      height: var(--height);
    }

    #fallback:hover {
      background-color: var(--magenta);
    }
  </style>
</head>

<body>
  <canvas>
    <div id="fallback"></div>
  </canvas>
</body>

<script>
  const canvas = document.querySelector("canvas");
  canvas.width = canvas.height = 200;
  const ctx = canvas.getContext("2d", { willReadFrequently: true });
  const fallbackDiv = document.getElementById("fallback");

  const style = getComputedStyle(document.body);
  const magenta = style.getPropertyValue('--magenta');
  const brightGreen = style.getPropertyValue('--brightGreen');
  const width = parseInt(style.getPropertyValue('--width'));
  const height = parseInt(style.getPropertyValue('--height'));

  function toRGB(imageData) {
    return `rgb(${imageData[0]}, ${imageData[1]}, ${imageData[2]})`
  }

  const t = async_test("placeElementHitTesting");
  async function testCorner(x, y, left, top) {
    async function movePointer(x, y) {
      const actions_promise = new test_driver.Actions()
        .pointerMove(x, y)
        .send();

      await actions_promise;
    }

    const x_outside = left ? x - 1 : x + 1;
    const y_outside = top ? y - 1 : y + 1;
    await movePointer(x_outside, y);
    t.step(() => {
      // The element is bright green when not hovered.
      const imageData = ctx.getImageData(x, y, 1, 1).data;
      assert_equals(toRGB(imageData), brightGreen);
    });
    await movePointer(x, y);
    t.step(() => {
      // Now that we're hovering, it should be magenta.

      const imageData = ctx.getImageData(x, y, 1, 1).data;
      assert_equals(toRGB(imageData), magenta);
    });
    await movePointer(x, y_outside);
    t.step(() => {
      const imageData = ctx.getImageData(x, y, 1, 1).data;
      assert_equals(toRGB(imageData), brightGreen);
    });
  }

  async function testPoint(x, y) {
    ctx.placeElement(fallbackDiv, x, y);

    // Upper left corner.

    await testCorner(x, y, true, true);
    // Upper right corner.
    await testCorner(x + width - 1, y, false, true);
    // Lower left corner.
    await testCorner(x, y + height - 1, true, false);
    // Lower right corner.
    await testCorner(x + width - 1, y + height - 1, false, false);
  }

  async function test() {
    await testPoint(100, 0);
    await testPoint(0, 0);
    await testPoint(0, 50);
    await testPoint(100, 50);
    t.done();
  }

  test();
</script>
